Give up on Paypal payout

+++ database/migrations/2017_03_15_103055_create_orders_table.php
@@ -15,7 +15,7 @@ class CreateOrdersTable extends Migration
     {
         Schema::create('orders', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('final_price')->unsigned()->nullable();
+            $table->float('final_price', 5, 2)->unsigned()->nullable();
             $table->integer('buyer_id')->unsigned();
             $table->string('deliver_location');
             $table->integer('deliverer_id')->unsigned()->nullable();
